+++ src/main/java/com/sarabarbara/compra/sheets/ItemSheet.java
@@ -7,6 +7,7 @@ import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.Size;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
@@ -45,7 +46,7 @@ public class ItemSheet {
 
     @DecimalMin(value = "0.01", message = "The price must be at least 0.01")
     @Column(name = "unit_price")
-    private float unitPrice;
+    private BigDecimal unitPrice;
 
     /**
      * The unitStock;
